`refactor`: melhor organização das exceções
