+++ backend/src/exceptions/base.py
@@ -1,0 +1,6 @@
+from fastapi import HTTPException
+
+
+class AppException(HTTPException):
+    def __init__(self, status_code: int, detail: str):
+        super().__init__(status_code=status_code, detail=detail)
